pyproject.toml
[tool.poetry]
name = "note-clerk"
version = "0.0.10"
authors = ["Adam Jackman <adam@acjackman.com>"]
description = "Command to manage plain text notes."
license = "MIT"
readme = "README.md"
homepage = "https://github.com/<your-username>/hypermodern-python"
repository = "https://github.com/<your-username>/hypermodern-python"
documentation = "https://note-clerk.readthedocs.io"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.0"
requests = "^2.23.0"
desert = "^2020.1.6"
marshmallow = "^3.5.1"
boltons = "^20.0.0"
python-frontmatter = "^0.5.0"
"ruamel.yaml" = "^0.16.12"
orderedset = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-mock = "^2.0.0"
coverage = {extras = ["toml"], version = "^5.0.3"}
pytest-cov = "^2.8.1"
black = "^20.8b1"
flake8 = "^3.7.9"
flake8-bandit = "^2.1.2"
flake8-black = "^0.1.1"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.8.5"
mypy = "^0.770"
# pytype = {version = "^2020.2.20", python = "3.7"}
typeguard = "^2.7.1"
flake8-annotations = "^2.0.1"
flake8-docstrings = "^1.5.0"
darglint = "^1.2.0"
xdoctest = "^0.11.0"
sphinx = "^3.0.0"
sphinx-autodoc-typehints = "^1.10.3"
codecov = "^2.0.16"

[tool.poetry.scripts]
note-clerk = "note_clerk.console:cli"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["note_clerk"]

[tool.coverage.report]
show_missing = true
fail_under = 100
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
